+++ src/cobra/sampling/__init__.py
@@ -1,4 +1,5 @@
-from .hr_sampler import HRSampler, shared_np_array, step
+from .hr_sampler import HRSampler, shared_np_array
 from .achr import ACHRSampler
+from .core import step
 from .optgp import OptGPSampler
 from .sampling import sample
